pyproject.toml
[tool.poetry]
name = "slaqbot"
version = "0.1.0"
description = ""
authors = ["Mark Lechner"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
slack-bolt = "^1.18.1"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
openai = "^1.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
